Only enable sqlite driver on linux because it requires cgo and fails to cross compile to other operating systems
